Day 2: SparkSeed progress
Today I focus in start growing the engine to the porpouse of the game I have in mind, while the game it's not a visual novel, this is a alpha version of the engine so for now I've focus in creating a simple visual novel. I learn how to add images as sprites or "characters" into my game. Also I manage to add more information about compilation and libraries in the README.txt, because I found some travel today while trying to compile in the other computer.